@@ -1,20 +1,17 @@
 source 'https://rubygems.org'
 
+gem 'bootsnap', require: false
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '~> 4.2.11'
+gem 'rails', '~> 6.0.4'
 
 # Use sqlite3 or mysql2 as the databases for Active Record
-gem 'sqlite3'
 gem 'mysql2', '~> 0.5'
 
 # Use SCSS for stylesheets
 gem 'sass-rails', '~> 5.0'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-gem 'coffee-rails', '~> 4.1.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
-gem 'therubyracer', '>= 0.12.3',  platforms: :ruby
 gem 'libv8', '>= 3.16.14.19' # Min version for Mac OS 10.11, XCode 9.0, Ruby 2.4
 
 # Use jquery as the JavaScript library
@@ -30,18 +27,17 @@ gem 'sdoc', '~> 0.4.0', group: :doc
 # Use Unicorn as the app server
 # gem 'unicorn'
 
-# Hydra stack
 gem 'nokogiri', '~> 1.8.1'
-gem 'blacklight', '~> 5.19'
-gem 'hydra-head', '~>7'
 
 # Columbia Hydra models
-gem 'cul_hydra', '~> 1.7.5'
+gem 'cul_hydra', :git => 'https://github.com/cul/cul_hydra.git', :branch => 'master'
 #gem 'cul_hydra', :path => '../cul_hydra'
 gem 'active_fedora_relsint', :git => 'https://github.com/cul/active_fedora_relsint.git', :branch => 'master'
-
+gem 'blacklight', '~> 7.22'
+gem 'view_component', '~> 2.51.0'
 # Use imogen for generating images
 gem 'imogen', '0.2.1'
+gem 'ruby-vips', '~> 2.0.16'
 #gem 'imogen', :path => '../imogen'
 
 # Use best type for mime/dc lookups
@@ -83,11 +79,12 @@ gem 'rack-protection', '>= 1.5.5'
 gem 'loofah', '>= 2.3.1'
 
 group :development, :test do
+  gem 'sqlite3'
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug'
 
-  gem 'rspec-rails', '~> 3.5'
-  gem 'jettywrapper', '>= 1.5.1'
+  gem 'rspec-rails', '~> 5.0.2'
+  gem 'jettywrapper', '>=2.0.5', git: 'https://github.com/samvera-deprecated/jettywrapper.git', branch: 'master'
 end
 
 group :development do
@@ -104,4 +101,5 @@ group :development do
   # The `deploy:restart` hook for passenger applications is now in a separate gem
   # Just add it to your Gemfile and require it in your Capfile.
   gem 'capistrano-passenger', '~> 0.1', require: false
+  gem 'listen'
 end
