@@ -1,3 +1,3 @@
-class DbCacheRecord < ActiveRecord::Base
+class DbCacheRecord < ApplicationRecord
   serialize :data, HashWithIndifferentAccess
 end
