@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,16 +10,15 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170106172259) do
+ActiveRecord::Schema.define(version: 2017_01_06_172259) do
 
   create_table "db_cache_records", force: :cascade do |t|
-    t.string   "pid"
-    t.text     "data"
-    t.boolean  "derivative_generation_in_progress", default: false, null: false
-    t.datetime "created_at",                                        null: false
-    t.datetime "updated_at",                                        null: false
+    t.string "pid"
+    t.text "data"
+    t.boolean "derivative_generation_in_progress", default: false, null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["pid"], name: "index_db_cache_records_on_pid", unique: true
   end
-
-  add_index "db_cache_records", ["pid"], name: "index_db_cache_records_on_pid", unique: true
 
 end
